+++ src/helpers/constants/Role.js
@@ -3,7 +3,7 @@ exports.sideNames = ["villagers", "werewolves"];
 exports.groupNames = [...this.sideNames, "lovers", "charmed"];
 
 exports.roles = [
-    { name: "werewolf", side: "werewolves", maxInGame: 4 },
+    { name: "werewolf", side: "werewolves", maxInGame: 39 },
     { name: "big-bad-wolf", side: "werewolves", maxInGame: 1, recommendedMinPlayers: 15 },
     { name: "vile-father-of-wolves", side: "werewolves", maxInGame: 1, recommendedMinPlayers: 12 },
     { name: "villager", side: "villagers", maxInGame: 39 },